Add batchTraverse usage example
I was very pleased to find this function to process Promises/Tasks in strict sequence (to avoid violating a rate limit on a API i am using), but i had difficulty to use it at first. Hopefully this example makes it easier for the next person that struggles with the same problem.